pyproject.toml
[project]
name = "sdkv1featureusagescan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
"adal==1.2.7",
"applicationinsights==0.11.10",
"argcomplete==3.6.3",
"attrs==25.4.0",
"azure-common==1.1.28",
"azure-core==1.36.0",
"azure-graphrbac==0.61.2",
"azure-identity==1.25.1",
"azure-mgmt-authorization==4.0.0",
"azure-mgmt-containerregistry==13.0.0",
"azure-mgmt-core==1.6.0",
"azure-mgmt-keyvault==11.0.0",
"azure-mgmt-network==29.0.0",
"azure-mgmt-resource==24.0.0",
"azure-mgmt-storage==23.0.0",
"azureml-automl-core==1.60.0",
"azureml-core==1.60.0.post1",
"azureml-datadrift==1.60.0",
"azureml-dataprep==5.1.6",
"azureml-dataprep-native==41.0.0",
"azureml-dataprep-rslex==2.22.5",
"azureml-dataset-runtime==1.60.0",
"azureml-pipeline==1.60.0",
"azureml-pipeline-core==1.60.0",
"azureml-pipeline-steps==1.60.0",
"azureml-sdk==1.60.0",
"azureml-telemetry==1.60.0",
"azureml-train-automl-client==1.60.0",
"azureml-train-core==1.60.0",
"azureml-train-restclients-hyperdrive==1.60.0",
"backports-tempfile==1.0",
"backports-weakref==1.0.post1",
"bcrypt==5.0.0",
"certifi==2025.10.5",
"cffi==2.0.0",
"charset-normalizer==3.4.4",
"cloudpickle==2.2.1",
"contextlib2==21.6.0",
"contourpy==1.3.2",
"cryptography==46.0.3",
"cycler==0.12.1",
"docker==7.1.0",
"fonttools==4.60.1",
"fusepy==3.0.1",
"humanfriendly==10.0",
"idna==3.11",
"importlib-metadata==8.2.0",
"importlib-resources==6.4.0",
"isodate==0.7.2",
"jeepney==0.9.0",
"jmespath==1.0.1",
"joblib==1.5.2",
"jsonpickle==4.1.1",
"jsonschema==4.25.1",
"jsonschema-specifications==2025.9.1",
"kiwisolver==1.4.9",
"knack==0.12.0",
"lightgbm==4.6.0",
"matplotlib==3.6.3",
"msal==1.34.0",
"msal-extensions==1.3.1",
"msrest==0.7.1",
"msrestazure==0.6.4.post1",
"ndg-httpsclient==0.5.1",
"numpy==1.23.5",
"oauthlib==3.3.1",
"packaging==25.0",
"pandas==1.5.3",
"paramiko==3.5.1",
"pathspec==0.12.1",
"pillow==12.0.0",
"pkginfo==1.12.1.2",
"py4j==0.10.9.9",
"pyarrow==22.0.0",
"pyasn1==0.6.1",
"pycparser==2.23",
"pygments==2.19.2",
"pyjwt==2.10.1",
"pynacl==1.6.0",
"pyopenssl==25.3.0",
"pyparsing==3.2.5",
"pyreadline3==3.5.4",
"pysocks==1.7.1",
"pyspark==4.0.1",
"python-dateutil==2.9.0.post0",
"pytz==2025.2",
"pywin32==311",
"pyyaml==6.0.3",
"referencing==0.37.0",
"requests==2.32.5",
"requests-oauthlib==2.0.0",
"rpds-py==0.28.0",
"scikit-learn==1.7.2",
"scipy==1.15.3",
"secretstorage==3.4.0",
"setuptools==80.9.0",
"six==1.17.0",
"tabulate==0.9.0",
"threadpoolctl==3.6.0",
"typing-extensions==4.15.0",
"urllib3==2.5.0",
"zipp==3.23.0"
]
